official experiments

+++ pure_sr.py
@@ -267,7 +267,7 @@ def main(
     try:
         num_cpus = int(os.environ.get('SLURM_CPUS_ON_NODE')) * int(os.environ.get('SLURM_JOB_NUM_NODES'))
     except TypeError:
-        num_cpus = 10
+        num_cpus = 5
 
     config = {
         'equation_file': f'sr_results/{version}.csv',
@@ -332,11 +332,12 @@ def main(
         traceback.print_exc()
 
     finally:
-        try:
-            # delete julia files: julia-1911988-17110333239-0016.out
-            subprocess.run(f'rm julia*.out', shell=True, check=True)
-        except subprocess.CalledProcessError as e:
-            pass
+        if os.path.exists('julia*.out'):
+            try:
+                # delete julia files: julia-1911988-17110333239-0016.out
+                subprocess.run(f'rm julia*.out', shell=True, check=True)
+            except subprocess.CalledProcessError as e:
+                print(f"An error occurred while trying to delete the backup files: {e}")
 
 
 if __name__ == "__main__":
